requirements.txt
Django>=4.2
Pillow>=10.0
